Assignment_7/pyproject.toml
[project]
name = "assignment-7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.10",
    "tavily-python",
]
